submit score as int, firewall always on top

+++ android/src/com/lappard/android/logic/GameData.java
@@ -9,5 +9,5 @@ public class GameData {
     public LevelData level;
     public long[] actions;
     public int date;
-    public float playerScore;
+    public int playerScore;
 }
